Fix type mismatch in getBoolean function
Changed value.asInt() != 0L to value.asInt().toInt() != 0 to fix Kotlin type mismatch error between Int and Long.

+++ app/src/main/java/com/phoenix/companionforcodblackops7/feature/tacticals/data/repository/TacticalsRepositoryImpl.kt
@@ -76,7 +76,7 @@ class TacticalsRepositoryImpl @Inject constructor(
             return when {
                 value == null -> default
                 value.type == RealmAny.Type.BOOL -> value.asBoolean()
-                value.type == RealmAny.Type.INT -> value.asInt() != 0L
+                value.type == RealmAny.Type.INT -> value.asInt().toInt() != 0
                 value.type == RealmAny.Type.STRING -> value.asString().equals("1", ignoreCase = true) || value.asString().equals("true", ignoreCase = true)
                 else -> default
             }
